+++ packages/experiment-tag/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.3](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.9.2...@amplitude/experiment-tag@0.9.3) (2025-08-29)
+
+
+### Bug Fixes
+
+* inject preview modal on applyVariants ([#209](https://github.com/amplitude/experiment-js-client/issues/209)) ([8348f68](https://github.com/amplitude/experiment-js-client/commit/8348f6874c61470d4f2f8270541c94c197a594dd))
+
+
+
+
+
+## [0.9.2](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.9.1...@amplitude/experiment-tag@0.9.2) (2025-08-28)
+
+
+### Bug Fixes
+
+* expire marketing cookie after session ([#208](https://github.com/amplitude/experiment-js-client/issues/208)) ([6bb8558](https://github.com/amplitude/experiment-js-client/commit/6bb85581030bb1130e83ca15e62d068bece9be43))
+
+
+
+
+
+## [0.9.1](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.9.0...@amplitude/experiment-tag@0.9.1) (2025-08-28)
+
+
+### Bug Fixes
+
+* add patch to removeChild to address hydration errors ([#207](https://github.com/amplitude/experiment-js-client/issues/207)) ([531d2f3](https://github.com/amplitude/experiment-js-client/commit/531d2f380e85849de498b016182ba11ff53bd249))
+* display modal during experiment preview mode ([6576834](https://github.com/amplitude/experiment-js-client/commit/657683437357cf32df24f509b947bc9c84034f12))
+* set marketing cookie on landing page before redirect ([#206](https://github.com/amplitude/experiment-js-client/issues/206)) ([cd2190a](https://github.com/amplitude/experiment-js-client/commit/cd2190ac34913ee411d5a2df6d553248158c214d))
+* Support preview mode persistence for MPAs ([#204](https://github.com/amplitude/experiment-js-client/issues/204)) ([14c35ee](https://github.com/amplitude/experiment-js-client/commit/14c35ee612c45f66b46bf0e1b0ff518ca4cdfa27))
+
+
+
+
+
+# [0.9.0](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.8.2...@amplitude/experiment-tag@0.9.0) (2025-08-11)
+
+
+### Features
+
+* Support visual editor mode for multi-page apps ([#198](https://github.com/amplitude/experiment-js-client/issues/198)) ([fb4310b](https://github.com/amplitude/experiment-js-client/commit/fb4310bd862fd9aadc917b84d390cb4bba9a39ca))
+
+
+
+
+
+## [0.8.2](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.8.1...@amplitude/experiment-tag@0.8.2) (2025-08-06)
+
+**Note:** Version bump only for package @amplitude/experiment-tag
+
+
+
+
+
 ## [0.8.1](https://github.com/amplitude/experiment-js-client/compare/@amplitude/experiment-tag@0.8.0...@amplitude/experiment-tag@0.8.1) (2025-07-28)
 
 **Note:** Version bump only for package @amplitude/experiment-tag